+++ backend/src/helpers/exceptions.py
@@ -1,6 +1,7 @@
 from fastapi import HTTPException
 
 
+#TODO: Create Words Exceptions
 INTERNAL_ERROR_EXCEPTION = HTTPException(
     status_code=500,
     detail="An internal error occured. Please, contact the support!"