+++ backend/src/models/models.py
@@ -1,4 +1,4 @@
 #File that alembic uses to get changes on the tables
 #================================#
-from ..models import users
+from ..models import users, words
 #================================#